+++ cmd/status.go
@@ -61,11 +61,8 @@ func runStatus(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("migrating config layout: %w", err)
 	}
 
-	instanceName := "default"
-	if len(args) > 0 {
-		instanceName = args[0]
-	}
-	if err := config.ValidateInstanceName(instanceName); err != nil {
+	instanceName, err := resolveOptionalInstanceName(args, "status", cmd.ErrOrStderr())
+	if err != nil {
 		return err
 	}
 	paths = paths.ForInstance(instanceName)