+++ cmd/stop.go
@@ -48,11 +48,8 @@ func runStop(cmd *cobra.Command, args []string) error {
 		return stopAllInstances(ctx, out, paths)
 	}
 
-	instanceName := "default"
-	if len(args) > 0 {
-		instanceName = args[0]
-	}
-	if err := config.ValidateInstanceName(instanceName); err != nil {
+	instanceName, err := resolveOptionalInstanceName(args, "stop", cmd.ErrOrStderr())
+	if err != nil {
 		return err
 	}